requirements.txt
Django==3.2
djangorestframework==3.12.4
psycopg2
dj-rest-auth
drf-yasg
django-cors-headers